+++ packages/nectar/src/sagas/publicChannels/publicChannels.selectors.test.ts
@@ -1,4 +1,4 @@
-import { keyFromCertificate, parseCertificate, setupCrypto } from '@zbayapp/identity'
+import { keyFromCertificate, parseCertificate, setupCrypto } from '@quiet/identity'
 import { Store } from '../store.types'
 import { getFactory, Identity, publicChannels } from '../..'
 import { prepareStore } from '../../utils/tests/prepareStore'
@@ -42,12 +42,12 @@ describe('publicChannelsSelectors', () => {
 
     alice = await factory.create<ReturnType<typeof identityActions.addNewIdentity>['payload']>(
       'Identity',
-      { id: community.id, zbayNickname: 'alice' }
+      { id: community.id, nickname: 'alice' }
     )
 
     john = await factory.create<ReturnType<typeof identityActions.addNewIdentity>['payload']>(
       'Identity',
-      { id: community.id, zbayNickname: 'john' }
+      { id: community.id, nickname: 'john' }
     )
 
     /* Messages ids are being used only for veryfing proper order...